Extract matrix axes to their own file
They have a separate section in `job.rb`, so they deserve their own
file rather than being grouped in with parameters.
