+++ lib/jujube/components.rb
@@ -1,5 +1,6 @@
 require_relative "components/macros"
 require_relative "components/helpers"
+require_relative "components/axes"
 require_relative "components/parameters"
 require_relative "components/properties"
 require_relative "components/scm"
@@ -13,6 +14,7 @@ module Jujube
   # Helper methods for creating jenkins-job-builder components.
   module Components
     include Helpers
+    include Axes
     include Parameters
     include Properties
     include Scm
